fixed bug [ 2010261 ] Java.flex example file can't handle most negative int literal
Now matching -2147483648 separately to return MIN_VALUE. 
Hex and Oct should still be handled with 2 tokens (MINUS literal).
An expression (x -2147483648) will be rejected, but should be, since 
"2147483648" on its own is not a valid literal.  Might be a bit harder to
detect this specific error case in parser and produce a good error message, 
but it makes the rest more uniform.


svn-rev: trunk@426