Fix issue #133 "Error in skeleton.nested"
The problem was that empty matches preceded `<<EOF>>` actions. This change set
pulls the EOF test including action and return code before the action switch
statement.

The drawback is that we now test for the infrequent EOF on every single action.

Since the issue can only occur if there are empty matches, an optimization
would be to only move the EOF test before the action switch if the spec
contains potential empty matches.

Since action code will usually dominate the EOF test anyway, I'm deferring this
optimization for now.

Test cases updated to the new skeleton.nested; the test case `genlook`
actually contained an instance of an empty match before EOF, but avoided
non-termination because it changed state on that match. The test case
`emptymatch` shows that empty matches within the stream still work.

Caveats:
  * Can't match the empty string at the end of the stream before EOF any
    more, because EOF will take precedence. Should be able to pull any such
    action into EOF code to avoid this problem.
  * `break` statements at the end of EOF code will cause a compile error now,
    because they are not in a switch statement any more. Solution: remove
    such `break` statements, they shouldn't be there in the first place.