add yychar reporting when both %debug and %char are present
The main method of %debug scanners will now report the value of yychar for
each match. As opposed to yyline and yycolumn counting who report yyline+1 and
yycolumn+1, the yychar reporting starts at 0, because it is more likely to be
useful as an index into a buffer (whereas line and column are more likely to
be useful in an editor).

Fixes issue #207.