Use a sparse two-level table for cmap lookup in generated scanner.
This addresses the memory issue #199. With more recent Unicode versions, the
character map took 0x110000 * 4 bytes, i.e. ~4MB. Most of this is never
accessed, but it still increases overall memory consumption esp for multiple
scanners.

This commit:
 * changes the current one-level array to a two-level table structure
 * adds quickcheck tests for the table construction
 * updates runtime engine to use that structure
 * updates skeleton files to remove reference to old ZZ_CMAP
 * removes trailing white space in generated template code

Typical memory consumption for the char map decreases from ~4MB to < 100KB,
for simple scanners (little unicode use > 0xFF) to ~20KB.

Even though this increases the number of operations in the innermost loop,
with a bit of luck performance might actually benefit because of better
cache locality. Benchmarking still to be done.