jflex: optional limit on token size
New directive

    %token_size_limit <limit>

where <limit> is a Java number literal or qualified identifier provides
an optional hard limit on token length.

The scanner buffer will not be increased beyond this length limit and
if a longer token is encountered the scanner with throw an EOFException.

This is for applications that require memory limits when parsing
untrusted input.

Implements #197