RegExp: establish IntCharSet.sub preconditions
`IntCharSet.sub(s)` expects `s` to be fully contained in `this`. If the
contents of the inner char class expression overlap, this assumption is
violated and leads to an inconsistent IntCharSet state.

Fix this by computing the union of the inner expression first, and then
returning the complement. Since the only difference to the CCLASS case
is the complement at the end, the two cases can be merged.

Fixes #1065